refactor score modifiers functions out
into base vespa index class
