+++ src/marqo/core/constants.py
@@ -14,3 +14,11 @@ MARQO_DOC_HYBRID_LEXICAL_SCORE = '_lexical_score'
 
 MARQO_STRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION = '2.10.0'
 MARQO_UNSTRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION = '2.11.0'
+
+# For score modifiers
+QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_2_9 = 'marqo__mult_weights'
+QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_2_9 = 'marqo__add_weights'
+QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_LEXICAL = 'marqo__mult_weights_lexical'
+QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_LEXICAL = 'marqo__add_weights_lexical'
+QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_TENSOR = 'marqo__mult_weights_tensor'
+QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_TENSOR = 'marqo__add_weights_tensor'
