+++ src/marqo/core/models/__init__.py
@@ -1,2 +1,2 @@
 from .marqo_index import MarqoIndex, UnstructuredMarqoIndex
-from .marqo_query import MarqoQuery
+from .marqo_query import MarqoQuery, MarqoHybridQuery, MarqoTensorQuery, MarqoLexicalQuery
