+++ src/marqo/core/structured_vespa_index/common.py
@@ -1,5 +1,3 @@
-from marqo.core.models.marqo_index import DistanceMetric
-
 FIELD_ID = "marqo__id"
 
 FIELD_SCORE_MODIFIERS_2_8 = 'marqo__score_modifiers' # retain for backwards compatibility
@@ -17,12 +15,6 @@ RANK_PROFILE_EMBEDDING_SIMILARITY_MODIFIERS_2_9 = 'embedding_similarity_modifier
 # Note field names are also used as query inputs, so make sure these reserved names have a marqo__ prefix
 QUERY_INPUT_EMBEDDING = 'marqo__query_embedding'
 QUERY_INPUT_BM25_AGGREGATOR = 'marqo__bm25_aggregator'
-QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_2_9 = 'marqo__mult_weights'
-QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_2_9 = 'marqo__add_weights'
-QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_LEXICAL = 'marqo__mult_weights_lexical'
-QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_LEXICAL = 'marqo__add_weights_lexical'
-QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_TENSOR = 'marqo__mult_weights_tensor'
-QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_TENSOR = 'marqo__add_weights_tensor'
 
 # For hybrid search
 RANK_PROFILE_HYBRID_CUSTOM_SEARCHER = 'hybrid_custom_searcher'
