+++ src/marqo/core/structured_vespa_index/structured_vespa_index.py
@@ -790,122 +790,6 @@ class StructuredVespaIndex(VespaIndex):
         else:
             return '*'
 
-    def _get_score_modifiers(self, marqo_query: MarqoQuery) -> Optional[Dict[str, Dict[str, float]]]:
-        """
-        Returns classic score modifiers (from tensor or lexical queries) as a dictionary of dictionaries.
-        Split between 'mult' and 'add' weights.
-        """
-        if marqo_query.score_modifiers:
-            mult_tensor = {}
-            add_tensor = {}
-            for modifier in marqo_query.score_modifiers:
-                if modifier.type == ScoreModifierType.Multiply:
-                    mult_tensor[modifier.field] = modifier.weight
-                elif modifier.type == ScoreModifierType.Add:
-                    add_tensor[modifier.field] = modifier.weight
-                else:
-                    raise InternalError(f'Unknown score modifier type {modifier.type}')
-
-            if self._marqo_index_version < self._HYBRID_SEARCH_MINIMUM_VERSION:
-                return {
-                    common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_2_9: mult_tensor,
-                    common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_2_9: add_tensor
-                }
-            elif isinstance(marqo_query, MarqoTensorQuery):
-                return {
-                    common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_TENSOR: mult_tensor,
-                    common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_TENSOR: add_tensor
-                }
-            elif isinstance(marqo_query, MarqoLexicalQuery):
-                return {
-                    common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_LEXICAL: mult_tensor,
-                    common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_LEXICAL: add_tensor
-                }
-            else:
-                raise InternalError(f'Unknown query type {type(marqo_query)}')
-
-        return None
-
-    def _get_hybrid_score_modifiers(self, hybrid_query: MarqoHybridQuery) -> \
-            Optional[Dict[str, Dict[str, Dict[str, float]]]]:
-
-        """
-        Specifically for hybrid queries.
-        Returns a dictionary with 2 keys: 'lexical' and 'tensor'.
-        Each key points to a dictionary containing the score modifiers for the respective field types.
-
-        Example:
-        {
-            'lexical': {
-                'marqo__mult_weights_lexical': {
-                    'field1': 0.5, 'field2': 0.4
-                },
-                'marqo__add_weights_lexical': {
-                    'field3': 23, 'field4': 12
-                }
-            },
-            'tensor': {
-                'marqo__mult_weights_tensor': {
-                    'field5': 0.5, 'field6': 0.4
-                },
-                'marqo__add_weights_tensor': {
-                    'field7': 23, 'field8': 12
-                }
-            }
-        }
-        """
-
-        result = {
-            constants.MARQO_SEARCH_METHOD_LEXICAL: None,
-            constants.MARQO_SEARCH_METHOD_TENSOR: None
-        }
-
-        def _convert_score_modifiers_to_tensors(score_modifiers: List[ScoreModifier], search_type: str) -> Dict[
-            str, Dict[str, float]]:
-            """
-            Helper function that converts a list of score modifiers into a dictionary of dictionaries.
-            The dictionaries are split between 'mult' and 'add' weights.
-
-            search_type: 'lexical' or 'tensor'. It will be added to the end of the
-            dictionary key to be parsed out in custom searcher. This allows the 2 score modifiers lists to be used
-            separately in the custom searcher.
-            """
-            mult_tensor = {}
-            add_tensor = {}
-            for modifier in score_modifiers:
-                if modifier.type == ScoreModifierType.Multiply:
-                    mult_tensor[modifier.field] = modifier.weight
-                elif modifier.type == ScoreModifierType.Add:
-                    add_tensor[modifier.field] = modifier.weight
-                else:
-                    raise InternalError(f'Unknown score modifier type {modifier.type}')
-            if search_type == constants.MARQO_SEARCH_METHOD_LEXICAL:
-                return {
-                    f"{common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_LEXICAL}": mult_tensor,
-                    f"{common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_LEXICAL}": add_tensor
-                }
-            elif search_type == constants.MARQO_SEARCH_METHOD_TENSOR:
-                return {
-                    f"{common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_TENSOR}": mult_tensor,
-                    f"{common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_TENSOR}": add_tensor
-                }
-            else:
-                raise InternalError(f"Unknown search type {search_type}")
-
-        if hybrid_query.score_modifiers_lexical:
-            result[constants.MARQO_SEARCH_METHOD_LEXICAL] = _convert_score_modifiers_to_tensors(
-                hybrid_query.score_modifiers_lexical,
-                constants.MARQO_SEARCH_METHOD_LEXICAL
-            )
-
-        if hybrid_query.score_modifiers_tensor:
-            result[constants.MARQO_SEARCH_METHOD_TENSOR] = _convert_score_modifiers_to_tensors(
-                hybrid_query.score_modifiers_tensor,
-                constants.MARQO_SEARCH_METHOD_TENSOR
-            )
-
-        return result
-
     def _get_lexical_search_term(self, marqo_query: MarqoLexicalQuery) -> str:
         if isinstance(marqo_query, MarqoHybridQuery):
             score_modifiers = marqo_query.hybrid_parameters.scoreModifiersLexical
