+++ src/marqo/core/unstructured_vespa_index/unstructured_vespa_index.py
@@ -216,102 +216,6 @@ class UnstructuredVespaIndex(VespaIndex):
         if marqo_query.filter is not None:
             return tree_to_filter_string(marqo_query.filter.root)
 
-    def _convert_score_modifiers_to_tensors(self, score_modifiers: List[ScoreModifier]) -> Dict[
-        str, Dict[str, float]]:
-        """
-        Helper function that converts a list of score modifiers into 2 dictionaries:
-        These dictionaries are 'mult' and 'add' weights.
-        """
-        mult_tensor = {}
-        add_tensor = {}
-        for modifier in score_modifiers:
-            if modifier.type == ScoreModifierType.Multiply:
-                mult_tensor[modifier.field] = modifier.weight
-            elif modifier.type == ScoreModifierType.Add:
-                add_tensor[modifier.field] = modifier.weight
-            else:
-                raise InternalError(f'Unknown score modifier type {modifier.type}')
-
-        return mult_tensor, add_tensor
-
-    def _get_score_modifiers(self, marqo_query: MarqoQuery) -> Optional[Dict[str, Dict[str, float]]]:
-        """
-        Returns classic score modifiers (from tensor or lexical queries) as a dictionary of dictionaries.
-        Split between 'mult' and 'add' weights.
-        """
-        if marqo_query.score_modifiers:
-            mult_tensor, add_tensor = self._convert_score_modifiers_to_tensors(marqo_query.score_modifiers)
-
-            if self._marqo_index_version < self._HYBRID_SEARCH_MINIMUM_VERSION:
-                return {
-                    unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_2_9: mult_tensor,
-                    unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_2_9: add_tensor
-                }
-            elif isinstance(marqo_query, MarqoTensorQuery):
-                return {
-                    unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_TENSOR: mult_tensor,
-                    unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_TENSOR: add_tensor
-                }
-            elif isinstance(marqo_query, MarqoLexicalQuery):
-                return {
-                    unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_LEXICAL: mult_tensor,
-                    unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_LEXICAL: add_tensor
-                }
-            else:
-                raise InternalError(f'Unknown query type {type(marqo_query)}')
-
-        return None
-
-    def _get_hybrid_score_modifiers(self, hybrid_query: MarqoHybridQuery) -> \
-            Optional[Dict[str, Dict[str, Dict[str, float]]]]:
-
-        """
-        Specifically for hybrid queries.
-        Returns a dictionary with 2 keys: 'lexical' and 'tensor'.
-        Each key points to a dictionary containing the score modifiers for the respective field types.
-
-        Example:
-        {
-            'lexical': {
-                'marqo__mult_weights_lexical': {
-                    'field1': 0.5, 'field2': 0.4
-                },
-                'marqo__add_weights_lexical': {
-                    'field3': 23, 'field4': 12
-                }
-            },
-            'tensor': {
-                'marqo__mult_weights_tensor': {
-                    'field5': 0.5, 'field6': 0.4
-                },
-                'marqo__add_weights_tensor': {
-                    'field7': 23, 'field8': 12
-                }
-            }
-        }
-        """
-
-        result = {
-            constants.MARQO_SEARCH_METHOD_LEXICAL: None,
-            constants.MARQO_SEARCH_METHOD_TENSOR: None
-        }
-
-        if hybrid_query.score_modifiers_lexical:
-            mult_tensor, add_tensor = self._convert_score_modifiers_to_tensors(hybrid_query.score_modifiers_lexical)
-            result[constants.MARQO_SEARCH_METHOD_LEXICAL] = {
-                unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_LEXICAL: mult_tensor,
-                unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_LEXICAL: add_tensor
-            }
-
-        if hybrid_query.score_modifiers_tensor:
-            mult_tensor, add_tensor = self._convert_score_modifiers_to_tensors(hybrid_query.score_modifiers_tensor)
-            result[constants.MARQO_SEARCH_METHOD_TENSOR] = {
-                unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_MULT_WEIGHTS_TENSOR: mult_tensor,
-                unstructured_common.QUERY_INPUT_SCORE_MODIFIERS_ADD_WEIGHTS_TENSOR: add_tensor
-            }
-
-        return result
-
     def _get_lexical_search_term(self, marqo_query: MarqoLexicalQuery) -> str:
         if isinstance(marqo_query, MarqoHybridQuery):
             score_modifiers = marqo_query.hybrid_parameters.scoreModifiersLexical
